requirements.txt
# If you don't have a GPU, you can run
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# Before installing the requirement.txt
torch
torchvision

anthropic
black[jupyter]
einops
inspect-ai==0.3.25
ipykernel
ipywidgets
jaxtyping
jupyter
matplotlib
notebook
openai
pandas
plotly
pre-commit
pytest
rich
scikit-learn
toml
tqdm
transformer_lens
transformer_utils
transformers
typeguard
typer
circuitsvis
opencv-python

InquirerPy
google-api-python-client 
google-auth 
google-auth-oauthlib 
google-auth-httplib2

# accelerate
# black
# click
# diffusers
# fancy_einsum
# flask
# ftfy
# gym==0.23.1
# joblib
# mediapy
# numpy-stl
# pygame
# requests
# sentence-transformers
# sympy
# tensorboard
# torchtyping
# transformers
# wandb
